refactor(web): cleaner code and better names for the tests
