+++ web/src/lib/components/shared-components/map/__tests__/clusters-integration.spec.ts
